Improve graphql controller structure
