+++ src/main/java/org/entur/lamassu/graphql/controller/GeofencingZonesGraphQLController.java
@@ -4,8 +4,8 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 import org.entur.lamassu.cache.GeofencingZonesCache;
+import org.entur.lamassu.graphql.validation.GraphQLQueryValidationService;
 import org.entur.lamassu.model.entities.GeofencingZones;
-import org.entur.lamassu.service.FeedProviderService;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
@@ -14,18 +14,19 @@ import org.springframework.stereotype.Controller;
 public class GeofencingZonesGraphQLController extends BaseGraphQLController {
 
   private final GeofencingZonesCache geofencingZonesCache;
+  private final GraphQLQueryValidationService validationService;
 
   public GeofencingZonesGraphQLController(
-    FeedProviderService feedProviderService,
-    GeofencingZonesCache geofencingZonesCache
+    GeofencingZonesCache geofencingZonesCache,
+    GraphQLQueryValidationService validationService
   ) {
-    super(feedProviderService);
     this.geofencingZonesCache = geofencingZonesCache;
+    this.validationService = validationService;
   }
 
   @QueryMapping
   public Collection<GeofencingZones> geofencingZones(@Argument List<String> systemIds) {
-    validateSystems(systemIds);
+    validationService.validateSystems(systemIds);
     if (systemIds != null && !systemIds.isEmpty()) {
       return geofencingZonesCache.getAll(Set.copyOf(systemIds));
     }
