+++ src/main/java/org/entur/lamassu/graphql/controller/StationGraphQLController.java
@@ -1,33 +1,16 @@
 package org.entur.lamassu.graphql.controller;
 
 import java.util.Collection;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import org.entur.lamassu.cache.PricingPlanCache;
-import org.entur.lamassu.cache.RegionCache;
 import org.entur.lamassu.cache.StationCache;
-import org.entur.lamassu.cache.SystemCache;
-import org.entur.lamassu.cache.VehicleTypeCache;
+import org.entur.lamassu.graphql.validation.GraphQLQueryValidationService;
 import org.entur.lamassu.model.entities.FormFactor;
-import org.entur.lamassu.model.entities.PricingPlan;
 import org.entur.lamassu.model.entities.PropulsionType;
-import org.entur.lamassu.model.entities.Region;
 import org.entur.lamassu.model.entities.Station;
-import org.entur.lamassu.model.entities.System;
-import org.entur.lamassu.model.entities.VehicleDocksAvailability;
-import org.entur.lamassu.model.entities.VehicleDocksCapacity;
-import org.entur.lamassu.model.entities.VehicleType;
-import org.entur.lamassu.model.entities.VehicleTypeAvailability;
-import org.entur.lamassu.model.entities.VehicleTypeCapacity;
-import org.entur.lamassu.model.entities.VehicleTypesCapacity;
 import org.entur.lamassu.service.*;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
-import org.springframework.graphql.data.method.annotation.SchemaMapping;
 import org.springframework.stereotype.Controller;
 
 @Controller
@@ -35,27 +18,16 @@ public class StationGraphQLController extends BaseGraphQLController {
 
   private final GeoSearchService geoSearchService;
   private final StationCache stationCache;
-  private final VehicleTypeCache vehicleTypeCache;
-  private final SystemCache systemCache;
-  private final PricingPlanCache pricingPlanCache;
-  private final RegionCache regionCache;
+  private final GraphQLQueryValidationService validationService;
 
   public StationGraphQLController(
     GeoSearchService geoSearchService,
-    FeedProviderService feedProviderService,
     StationCache stationCache,
-    VehicleTypeCache vehicleTypeCache,
-    SystemCache systemCache,
-    PricingPlanCache pricingPlanCache,
-    RegionCache regionCache
+    GraphQLQueryValidationService validationService
   ) {
-    super(feedProviderService);
     this.geoSearchService = geoSearchService;
     this.stationCache = stationCache;
-    this.vehicleTypeCache = vehicleTypeCache;
-    this.systemCache = systemCache;
-    this.pricingPlanCache = pricingPlanCache;
-    this.regionCache = regionCache;
+    this.validationService = validationService;
   }
 
   @QueryMapping
@@ -84,9 +56,9 @@ public class StationGraphQLController extends BaseGraphQLController {
       return stationCache.getAll(Set.copyOf(ids));
     }
 
-    validateCount(count);
-    validateCodespaces(codespaces);
-    validateSystems(systems);
+    validationService.validateCount(count);
+    validationService.validateCodespaces(codespaces);
+    validationService.validateSystems(systems);
 
     var filterParams = new StationFilterParameters(
       codespaces,
@@ -99,7 +71,7 @@ public class StationGraphQLController extends BaseGraphQLController {
 
     Collection<Station> stations;
 
-    validateQueryParameters(
+    validationService.validateQueryParameters(
       lat,
       lon,
       range,
@@ -109,7 +81,7 @@ public class StationGraphQLController extends BaseGraphQLController {
       maximumLongitude
     );
 
-    if (isRangeSearch(range, lat, lon)) {
+    if (validationService.isRangeSearch(range, lat, lon)) {
       var queryParams = new RangeQueryParameters(lat, lon, range);
       stations = geoSearchService.getStationsWithinRange(queryParams, filterParams);
     } else {
@@ -128,114 +100,5 @@ public class StationGraphQLController extends BaseGraphQLController {
   @QueryMapping
   public Collection<Station> stationsById(@Argument List<String> ids) {
     return stationCache.getAll(Set.copyOf(ids));
-  }
-
-  @SchemaMapping(typeName = "VehicleTypeAvailability", field = "vehicleType")
-  public VehicleType getVehicleType(VehicleTypeAvailability vehicleTypeAvailability) {
-    return vehicleTypeCache.get(vehicleTypeAvailability.getVehicleTypeId());
-  }
-
-  @SchemaMapping(typeName = "VehicleDocksAvailability", field = "vehicleTypes")
-  public List<VehicleType> getVehicleType(
-    VehicleDocksAvailability vehicleDocksAvailability
-  ) {
-    return vehicleTypeCache.getAll(
-      new HashSet<>(vehicleDocksAvailability.getVehicleTypeIds())
-    );
-  }
-
-  @SchemaMapping(typeName = "VehicleTypeCapacity", field = "vehicleType")
-  public VehicleType getVehicleType(VehicleTypeCapacity vehicleTypeCapacity) {
-    return vehicleTypeCache.get(vehicleTypeCapacity.getVehicleTypeId());
-  }
-
-  @SchemaMapping(typeName = "VehicleTypesCapacity", field = "vehicleTypes")
-  public List<VehicleType> getVehicleType(VehicleTypesCapacity vehicleTypesCapacity) {
-    return vehicleTypeCache.getAll(
-      new HashSet<>(vehicleTypesCapacity.getVehicleTypeIds())
-    );
-  }
-
-  @SchemaMapping(typeName = "VehicleDocksCapacity", field = "vehicleTypes")
-  public List<VehicleType> getVehicleType(VehicleDocksCapacity vehicleDocksCapacity) {
-    return vehicleTypeCache.getAll(
-      new HashSet<>(vehicleDocksCapacity.getVehicleTypeIds())
-    );
-  }
-
-  @SchemaMapping(typeName = "Station", field = "system")
-  public System getSystem(Station station) {
-    return systemCache.get(station.getSystemId());
-  }
-
-  /**
-   * GBFS does not have pricing plans directly on station. They should be resolved
-   * via vehicle types instead. This is a workaround for not having to resolve
-   * all of a system's pricing plans, by collecting only the pricing plan's referred
-   * to by a stations various references to vehicle types
-   */
-  @SchemaMapping(typeName = "Station", field = "pricingPlans")
-  public List<PricingPlan> getPricingPlans(Station station) {
-    Set<String> vehicleTypeIds = Stream
-      .of(
-        Optional
-          .ofNullable(station.getVehicleCapacity())
-          .orElse(List.of())
-          .stream()
-          .map(VehicleTypeCapacity::getVehicleTypeId),
-        Optional
-          .ofNullable(station.getVehicleDocksCapacity())
-          .orElse(List.of())
-          .stream()
-          .map(VehicleDocksCapacity::getVehicleTypeIds)
-          .flatMap(Collection::stream),
-        Optional
-          .ofNullable(station.getVehicleTypeCapacity())
-          .orElse(List.of())
-          .stream()
-          .map(VehicleTypeCapacity::getVehicleTypeId),
-        Optional
-          .ofNullable(station.getVehicleTypesCapacity())
-          .orElse(List.of())
-          .stream()
-          .map(VehicleTypesCapacity::getVehicleTypeIds)
-          .flatMap(Collection::stream),
-        Optional
-          .ofNullable(station.getVehicleTypesAvailable())
-          .orElse(List.of())
-          .stream()
-          .map(VehicleTypeAvailability::getVehicleTypeId),
-        Optional
-          .ofNullable(station.getVehicleDocksAvailable())
-          .orElse(List.of())
-          .stream()
-          .map(VehicleDocksAvailability::getVehicleTypeIds)
-          .flatMap(Collection::stream)
-      )
-      .flatMap(i -> i)
-      .collect(Collectors.toSet());
-
-    List<VehicleType> vehicleTypes = vehicleTypeCache.getAll(vehicleTypeIds);
-
-    Set<String> pricingPlanIds = new HashSet<>();
-
-    vehicleTypes.forEach(vehicleType -> {
-      if (vehicleType.getPricingPlanIds() != null) {
-        pricingPlanIds.addAll(vehicleType.getPricingPlanIds());
-      }
-      if (vehicleType.getDefaultPricingPlanId() != null) {
-        pricingPlanIds.add(vehicleType.getDefaultPricingPlanId());
-      }
-    });
-
-    return pricingPlanCache.getAll(pricingPlanIds);
-  }
-
-  @SchemaMapping(typeName = "Station", field = "region")
-  public Region getRegion(Station station) {
-    if (station.getRegionId() == null) {
-      return null;
-    }
-    return regionCache.get(station.getRegionId());
   }
 }
