+++ src/main/java/org/entur/lamassu/graphql/controller/VehicleGraphQLController.java
@@ -3,24 +3,17 @@ package org.entur.lamassu.graphql.controller;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
-import org.entur.lamassu.cache.PricingPlanCache;
-import org.entur.lamassu.cache.SystemCache;
 import org.entur.lamassu.cache.VehicleCache;
-import org.entur.lamassu.cache.VehicleTypeCache;
+import org.entur.lamassu.graphql.validation.GraphQLQueryValidationService;
 import org.entur.lamassu.model.entities.FormFactor;
-import org.entur.lamassu.model.entities.PricingPlan;
 import org.entur.lamassu.model.entities.PropulsionType;
-import org.entur.lamassu.model.entities.System;
 import org.entur.lamassu.model.entities.Vehicle;
-import org.entur.lamassu.model.entities.VehicleType;
 import org.entur.lamassu.service.BoundingBoxQueryParameters;
-import org.entur.lamassu.service.FeedProviderService;
 import org.entur.lamassu.service.GeoSearchService;
 import org.entur.lamassu.service.RangeQueryParameters;
 import org.entur.lamassu.service.VehicleFilterParameters;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
-import org.springframework.graphql.data.method.annotation.SchemaMapping;
 import org.springframework.stereotype.Controller;
 
 @Controller
@@ -28,24 +21,16 @@ public class VehicleGraphQLController extends BaseGraphQLController {
 
   private final GeoSearchService geoSearchService;
   private final VehicleCache vehicleCache;
-  private final VehicleTypeCache vehicleTypeCache;
-  private final PricingPlanCache pricingPlanCache;
-  private final SystemCache systemCache;
+  private final GraphQLQueryValidationService validationService;
 
   public VehicleGraphQLController(
     GeoSearchService geoSearchService,
-    FeedProviderService feedProviderService,
     VehicleCache vehicleCache,
-    VehicleTypeCache vehicleTypeCache,
-    PricingPlanCache pricingPlanCache,
-    SystemCache systemCache
+    GraphQLQueryValidationService validationService
   ) {
-    super(feedProviderService);
     this.geoSearchService = geoSearchService;
     this.vehicleCache = vehicleCache;
-    this.vehicleTypeCache = vehicleTypeCache;
-    this.pricingPlanCache = pricingPlanCache;
-    this.systemCache = systemCache;
+    this.validationService = validationService;
   }
 
   @QueryMapping
@@ -76,9 +61,9 @@ public class VehicleGraphQLController extends BaseGraphQLController {
       return vehicleCache.getAll(Set.copyOf(ids));
     }
 
-    validateCount(count);
-    validateCodespaces(codespaces);
-    validateSystems(systems);
+    validationService.validateCount(count);
+    validationService.validateCodespaces(codespaces);
+    validationService.validateSystems(systems);
 
     var filterParams = new VehicleFilterParameters(
       codespaces,
@@ -93,7 +78,7 @@ public class VehicleGraphQLController extends BaseGraphQLController {
 
     Collection<Vehicle> vehicles;
 
-    validateQueryParameters(
+    validationService.validateQueryParameters(
       lat,
       lon,
       range,
@@ -103,7 +88,7 @@ public class VehicleGraphQLController extends BaseGraphQLController {
       maximumLongitude
     );
 
-    if (isRangeSearch(range, lat, lon)) {
+    if (validationService.isRangeSearch(range, lat, lon)) {
       var queryParams = new RangeQueryParameters(lat, lon, range);
       vehicles = geoSearchService.getVehiclesWithinRange(queryParams, filterParams);
     } else {
@@ -117,20 +102,5 @@ public class VehicleGraphQLController extends BaseGraphQLController {
     }
 
     return vehicles;
-  }
-
-  @SchemaMapping(typeName = "Vehicle", field = "vehicleType")
-  public VehicleType vehicleType(Vehicle vehicle) {
-    return vehicleTypeCache.get(vehicle.getVehicleTypeId());
-  }
-
-  @SchemaMapping(typeName = "Vehicle", field = "pricingPlan")
-  public PricingPlan pricingPlan(Vehicle vehicle) {
-    return pricingPlanCache.get(vehicle.getPricingPlanId());
-  }
-
-  @SchemaMapping(typeName = "Vehicle", field = "system")
-  public System system(Vehicle vehicle) {
-    return systemCache.get(vehicle.getSystemId());
   }
 }
